+++ src/Pivot.php
@@ -12,11 +12,11 @@ class Pivot
     protected $table;
 
     /**
-     * Modèles concernés par la liaison pivot, par clé.
+     * Contient les instances des deux modèles concernés par la liaison pivot.
      *
      * @var array[Model]
      */
-    protected $models;
+    protected $relatedModels;
 
     /**
      * Constructeur.
@@ -29,42 +29,43 @@ class Pivot
     public function __construct($table, $foreignKey1 = null, $foreignKey2 = null)
     {
         $this->table = $table;
-        $this->models = [];
+        $this->relatedModels = [];
 
         if ($foreignKey1) {
-            $this->models[$foreignKey1] = null;
+            $this->relatedModels[$foreignKey1] = null;
         }
 
         if ($foreignKey2) {
-            $this->models[$foreignKey2] = null;
+            $this->relatedModels[$foreignKey2] = null;
         }
     }
 
     /**
-     * Ajoute/modifie le modèle associé à une clé.
+     * Ajoute/modifie le modèle concerné par la liaison pivot pour une des clés
+     * étrangères donnée.
      *
      * @param  string $foreignKey
-     * @param  Model  $model
+     * @param  Model  $relatedModels
      * @return void
      */
-    public function setModel($foreignKey, Model $model)
+    public function setRelatedModel($foreignKey, Model $relatedModels)
     {
-        $this->models[$foreignKey] = $model;
+        $this->relatedModels[$foreignKey] = $relatedModels;
     }
 
     /**
-     * Ajoute les relations n-n entre les 2 modèles concernés par le pivot,
-     * s'il y a bien 2 modèles de renseignés.
+     * Ajoute les relations n-n entre les deux modèles concernés par le pivot,
+     * s'il y a bien au moins deux modèles de renseignés.
      *
      * @return void
      */
-    public function addBelongsToManyRelationsToModels()
+    public function addBelongsToManyRelationsToRelatedModels()
     {
-        $foreignKeys = array_keys($this->models);
+        $foreignKeys = array_keys($this->relatedModels);
 
         if (count($foreignKeys) < 2
-            || empty($this->models[$foreignKeys[0]])
-            || empty($this->models[$foreignKeys[1]])) {
+            || empty($this->relatedModels[$foreignKeys[0]])
+            || empty($this->relatedModels[$foreignKeys[1]])) {
 
             throw new \Exception(
                 'Not enough keys in pivot table '.$this->table.' for creating many-to-many relations'
@@ -72,16 +73,16 @@ class Pivot
         }
 
         // Relation n-n du modèle A vers le modèle B
-        $this->models[$foreignKeys[0]]->belongsToMany(
-            $this->models[$foreignKeys[1]],
+        $this->relatedModels[$foreignKeys[0]]->belongsToMany(
+            $this->relatedModels[$foreignKeys[1]],
             $this->table,
             $foreignKeys[0],
             $foreignKeys[1]
         );
 
         // Relation n-n du modèle B vers le modèle A
-        $this->models[$foreignKeys[1]]->belongsToMany(
-            $this->models[$foreignKeys[0]],
+        $this->relatedModels[$foreignKeys[1]]->belongsToMany(
+            $this->relatedModels[$foreignKeys[0]],
             $this->table,
             $foreignKeys[1],
             $foreignKeys[0]