+++ src/Relations/BelongsToMany.php
@@ -105,7 +105,7 @@ class BelongsToMany extends Relation
     protected function buildName()
     {
         return camel_case($this->relatedModel->getTable())
-            . $this->getNamePrecision();
+             . $this->getNamePrecision();
     }
 
     /**